feat: fix pr coments
