+++ src/api/models/storage/StorageFilter.ts
@@ -16,8 +16,8 @@ export const mapToTransport = ({
   periods: Array.from(periods)
     .map((p: SubscriptionPeriod) => PeriodInSeconds[p]),
   averagePrice: {
-    min: price.min,
-    max: price.max,
+    min: price.min - 1,
+    max: price.max + 1,
   },
   totalCapacity: {
     min: sizeGB.min * UNIT_PREFIX_POW2.KILO,
