+++ src/context/App/AppContext.tsx
@@ -4,7 +4,8 @@ import { XRService } from 'api/rif-marketplace-cache/rates/xr'
 import { DomainsService } from 'api/rif-marketplace-cache/rns/domains'
 import { OffersService } from 'api/rif-marketplace-cache/rns/offers'
 import { SoldDomainsService } from 'api/rif-marketplace-cache/rns/sold'
-import { AvgBillingPriceService, StorageOffersService } from 'api/rif-marketplace-cache/storage/offers'
+import { AvgBillingPriceService } from 'api/rif-marketplace-cache/storage/avg-billing-plan-price'
+import { StorageOffersService } from 'api/rif-marketplace-cache/storage/offers'
 import { ContextActions, ContextReducer, ContextState } from 'context/storeUtils/interfaces'
 import storeReducerFactory from 'context/storeUtils/reducer'
 import {
